Follow / Unfollow User, Début des Regex pour les formulaires

+++ myapp/routes/index.js
@@ -13,7 +13,7 @@ const utils = require('./../config/utils') //generation Token
 const UserControl = require('./../controlers/usercontroler')
 const ThreadControl = require('./../controlers/threadControlers')
 const LikeControl = require('./../controlers/likeControler')
-
+const FollowControl = require('./../controlers/followControler')
 
 // HOME PAGE
 
@@ -33,29 +33,38 @@ router.get('/', isLoggedIn, function (req, res, next) {
     .then((data) => {
       ReadPushText(data)
         .then((Threads) => {
-          res.render('pages/index', { message: "message", threads: Threads, csurfToken: req.csrfToken() })
+          res.render('pages/index', { message: "message", threads: Threads, user_connect: req.session.passport.user, csurfToken: req.csrfToken() })
         })
     })
 
 })
 
 router.get('/thread?:hashtag', isLoggedIn, (req, res, next) => {
-  console.log(req.query.hashtag)
-
   ThreadControl.getThreadByHashtag(req.query.hashtag, req.session.passport.user)
     .then((data) => {
       ReadPushText(data)
         .then((Threads) => {
-          res.render('pages/index', { message: "message", threads: Threads, csurfToken: req.csrfToken(), hashtag: req.query.hashtag })
+          res.render('pages/index', { message: "message", threads: Threads, csurfToken: req.csrfToken(), hashtag: req.query.hashtag, user_connect: req.session.passport.user })
         })
     })
 })
 
 
 router.post('/thread/newthread', isLoggedIn, (req, res, next) => {
-  ThreadControl.newThreadFile(req.session.passport.user, req.body.title, req.body.thread_content)
-  ThreadControl.newThread(req.body.title, './thread/' + req.session.passport.user + '/' + req.body.title + '.html', "N", req.session.passport.user, req.body.hashtag)
-  res.redirect('/')
+  var regex = /^A-Za-z0-9/
+
+  var RegexResultTitle = regex.test(req.body.title)
+  var RegexResultContent = regex.test(req.body.thread_content)
+
+  if (RegexResultTitle === false && RegexResultContent === false) {
+    res.redirect('/profile')
+  } else {
+    ThreadControl.newThreadFile(req.session.passport.user, req.body.title, req.body.thread_content)
+    ThreadControl.newThread(req.body.title, './thread/' + req.session.passport.user + '/' + req.body.title + '.html', "N", req.session.passport.user, req.body.hashtag)
+    res.redirect('/')
+  }
+
+
 })
 
 router.get('/thread/delete?:id_thread', isLoggedIn, (req, res, next) => {
@@ -75,13 +84,23 @@ router.get('/thread/edit', isLoggedIn, (req, res, next) => {
 })
 
 router.post('/thread/edit', isLoggedIn, (req, res, next) => {
-  ThreadControl.getThreadById(req.body.id_thread)
-    .then((data) => {
-      ThreadControl.removeFile(data[0].pathfile_thread)
-      ThreadControl.newThreadFile(req.session.passport.user, req.body.title, req.body.thread_content)
-      ThreadControl.updateThread(req.body.id_thread, req.body.title, './thread/' + req.session.passport.user + '/' + req.body.title + '.html')
-      res.redirect('/')
-    })
+
+  var regex = /^A-Za-z0-9/
+
+  var RegexResultTitle = regex.test(req.body.title)
+  var RegexResultContent = regex.test(req.body.thread_content)
+
+  if (RegexResultTitle === false && RegexResultContent === false) {
+    res.redirect('/profile')
+  } else {
+    ThreadControl.getThreadById(req.body.id_thread)
+      .then((data) => { 
+        ThreadControl.removeFile(data[0].pathfile_thread)
+        ThreadControl.newThreadFile(req.session.passport.user, req.body.title, req.body.thread_content)
+        ThreadControl.updateThread(req.body.id_thread, req.body.title, './thread/' + req.session.passport.user + '/' + req.body.title + '.html', req.body.hashtag)
+        res.redirect('/')
+      })
+  }
 })
 
 router.get('/thread/random', isLoggedIn, (req, res, next) => {
@@ -89,10 +108,29 @@ router.get('/thread/random', isLoggedIn, (req, res, next) => {
     .then((data) => {
       ReadPushText(data)
         .then((Thread) => {
+          Thread[0].id_user_connect = req.session.passport.user
           res.json(Thread)
         })
     })
 })
+
+// FOLLOW BUTTON
+
+router.get('/follow', isLoggedIn, (req, res, next) => {
+  FollowControl.getFollow(req.query.id_user, req.session.passport.user)
+    .then((data) => {
+      if (data === false) {
+        FollowControl.FollowButton(req.query.id_user, req.session.passport.user)
+        bool = true
+        res.json(bool)
+      } else {
+        FollowControl.unFollowButton(req.query.id_user, req.session.passport.user)
+        bool = false
+        res.json(bool)
+      }
+    })
+})
+
 
 // LIKE BUTTON
 
@@ -148,12 +186,15 @@ router.get('/profile', isLoggedIn, function (req, res, next) {
     })
 })
 
+// PROFILES
+
 router.get('/profiles?:username', isLoggedIn, (req, res, next) => {
 
   const waitFor = (ms) => new Promise(r => setTimeout(r, ms));
 
-  UserControl.getUserByUsername(req.query.username)
+  UserControl.getUserByUsername(req.query.username, req.session.passport.user)
     .then((User) => {
+      console.log(User)
       if (User.id_user === req.session.passport.user) {
         res.redirect('/profile')
       }
@@ -161,7 +202,7 @@ router.get('/profiles?:username', isLoggedIn, (req, res, next) => {
         .then((data) => {
           ReadPushText(data)
             .then((Threads) => {
-              res.render('pages/profiles', { User: User, threads: Threads, csurfToken: req.csrfToken() })
+              res.render('pages/profiles', { User: User, threads: Threads, user_connect: req.session.passport.user, csurfToken: req.csrfToken() })
             })
         })
     })
